extracted Apollo client and loaders logic from App.js into separate files
