+++ client/src/utils/apolloClient.js
@@ -1,0 +1,32 @@
+import {
+    ApolloClient,
+    InMemoryCache,
+    createHttpLink,
+  } from '@apollo/client';
+import { setContext } from '@apollo/client/link/context';
+
+// Construct our main GraphQL API endpoint
+const httpLink = createHttpLink({
+    uri: (process.env.NODE_ENV === 'development') ? 'http://localhost:3001/graphql' : '/graphql',
+});
+  
+// Construct request middleware that will attach the JWT token to every request as an `authorization` header
+const authLink = setContext((_, { headers }) => {
+    // get the authentication token from local storage if it exists
+    const token = localStorage.getItem('id_token');
+    // return the headers to the context so httpLink can read them
+    return {
+        headers: {
+            ...headers,
+            authorization: token ? `Bearer ${token}` : '',
+        },
+    };
+});
+
+// Set up our client to execute the `authLink` middleware prior to making the request to our GraphQL API
+const client = new ApolloClient({
+    link: authLink.concat(httpLink),
+    cache: new InMemoryCache(),
+});
+
+export { client };
